Reestructuracion de componente DeleteIcon
